finished sending and receiving messages in chat.

+++ api/database/migrations/2019_07_29_211707_create_users_table.php
@@ -13,28 +13,29 @@ class CreateUsersTable extends Migration
      */
     public function up()
     {
-        Schema::create('users', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->bigInteger('role_id')->unsigned()->index();
-            $table->string('name');
-            $table->string('surname');
-            $table->date('birthday');
-            $table->string('email')->unique()->index();
-            $table->string('password');
-            $table->string('reset_password_token')->nullable();
-            $table->string('address');
-            $table->string('city');
-            $table->string('state');
-            $table->string('postal');
-            $table->string('phone_number');
-            $table->timestamps();
+        if (!Schema::hasTable('users'))
+            Schema::create('users', function (Blueprint $table) {
+                $table->bigIncrements('id');
+                $table->bigInteger('role_id')->unsigned()->index();
+                $table->string('name');
+                $table->string('surname');
+                $table->date('birthday');
+                $table->string('email')->unique()->index();
+                $table->string('password');
+                $table->string('reset_password_token')->nullable();
+                $table->string('address');
+                $table->string('city');
+                $table->string('state');
+                $table->string('postal');
+                $table->string('phone_number');
+                $table->timestamps();
 
-            $table->foreign('role_id')
-                ->references('id')
-                ->on('roles')
-                ->onDelete('restrict')
-                ->onUpdate('cascade');
-        });
+                $table->foreign('role_id')
+                    ->references('id')
+                    ->on('roles')
+                    ->onDelete('restrict')
+                    ->onUpdate('cascade');
+            });
     }
 
     /**
