+++ api/database/migrations/2019_07_29_212119_create_chats_table.php
@@ -13,25 +13,26 @@ class CreateChatsTable extends Migration
      */
     public function up()
     {
-        Schema::create('chats', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->bigInteger('sender')->unsigned()->index()->nullable();
-            $table->bigInteger('receiver')->unsigned()->index()->nullable();
-            $table->text('content');
-            $table->timestamps();
+        if (!Schema::hasTable('chats'))
+            Schema::create('chats', function (Blueprint $table) {
+                $table->bigIncrements('id');
+                $table->bigInteger('sender')->unsigned()->index()->nullable();
+                $table->bigInteger('receiver')->unsigned()->index()->nullable();
+                $table->text('content');
+                $table->timestamps();
 
-            $table->foreign('receiver')
-                ->references('id')
-                ->on('users')
-                ->onDelete('set null')
-                ->onUpdate('cascade');
-            
-            $table->foreign('sender')
-                ->references('id')
-                ->on('users')
-                ->onDelete('set null')
-                ->onUpdate('cascade');    
-        });
+                $table->foreign('receiver')
+                    ->references('id')
+                    ->on('users')
+                    ->onDelete('set null')
+                    ->onUpdate('cascade');
+
+                $table->foreign('sender')
+                    ->references('id')
+                    ->on('users')
+                    ->onDelete('set null')
+                    ->onUpdate('cascade');
+            });
     }
 
     /**
